[Static pages] Update documentation structure and rendering. Add a lot of new pages with first content
