+++ package.json
@@ -19,13 +19,13 @@
     "build:babel": "babel ./source/patterns --ignore *.md,*.txt,*.mustache,*.spec.js,*.json  --out-dir ./build/apparena-patterns-react/patterns --copy-files",
     "build:copy-files": "babel-node ./build/scripts/copy-files.js",
     "build:docs": "babel-node build/generator/index.js",
-    "build:frontend:dev": "npm run build:docs && cross-env BABEL_ENV=wildCard webpack --progress --config build/webpack/frontend/webpack.config.js",
-    "build:frontend:watch": "npm run build:docs && cross-env BABEL_ENV=wildCard webpack -w --progress --config build/webpack/frontend/webpack.config.js",
-    "build:frontend:prod": "npm run build:docs && cross-env BABEL_ENV=wildCard NODE_ENV=production webpack --progress --config build/webpack/frontend/webpack.config.js",
+    "build:frontend:dev": "npm run build:docs && webpack --progress --config build/webpack/frontend/webpack.config.js",
+    "build:frontend:watch": "npm run build:docs && webpack -w --progress --config build/webpack/frontend/webpack.config.js",
+    "build:frontend:prod": "npm run build:docs && cross-env NODE_ENV=production webpack --progress --config build/webpack/frontend/webpack.config.js",
     "clean:build:all": "rimraf build/apparena-patterns-react",
     "start": "npm run build:docs && node build/webpack/server.js",
     "test": "BABEL_DISABLE_CACHE=1 cross-env NODE_ENV=test nyc ava",
-    "test:generator": "cross-env BABEL_ENV=wildCard webpack --progress --config build/webpack/webpack.tests.js && BABEL_DISABLE_CACHE=1 cross-env NODE_ENV=test nyc ava build/generator/_build/*.js",
+    "test:generator": "webpack --progress --config build/webpack/webpack.tests.js && BABEL_DISABLE_CACHE=1 cross-env NODE_ENV=test nyc ava build/generator/_build/*.js",
     "test:coverage": "BABEL_DISABLE_CACHE=1 nyc report --reporter=lcov --reporter=json --reporter=text-lcov --reporter=html",
     "publish:generator:apparena-pattern": "npm publish build/yeoman/generator-apparena-pattern"
   },
