+++ source/pages/brand/index.js
@@ -1,0 +1,21 @@
+import * as Personality from './01-personality';
+import * as Colors from './colors';
+import * as Documents from './documents';
+import * as Icons from './icons';
+import * as Logo from './logos';
+import * as Presentations from './presentations';
+import * as Storytelling from './storytelling';
+import * as Typography from './typography';
+import * as WritingStyle from './writing-style';
+
+export {
+    Personality,
+    Colors,
+    Documents,
+    Icons,
+    Logo,
+    Presentations,
+    Storytelling,
+    Typography,
+    WritingStyle
+}
