--- source/pages/corporate-identity/styles.ci
@@ -1,73 +1,0 @@
-@import "~bootstrap/scss/functions";
-
-@import "../../css/scss/abstracts/variables";
-@import "../../css/scss/abstracts/mixins";
-@import "../../css/scss/base/colors";
-@import "../../css/scss/base/headings";
-
-/*------------------------------------*\
-    Vendor
-\*------------------------------------*/
-//@import "~bootstrap/scss/variables";
-//@import "~bootstrap/scss/mixins";
-@import "~bootstrap/scss/variables";
-@import "~bootstrap/scss/mixins";
-@import "~bootstrap/scss/print";
-@import "~bootstrap/scss/images";
-@import "~bootstrap/scss/code";
-@import "~bootstrap/scss/grid";
-@import "~bootstrap/scss/tables";
-@import "~bootstrap/scss/forms";
-@import "~bootstrap/scss/buttons";
-@import "~bootstrap/scss/transitions";
-@import "~bootstrap/scss/dropdown";
-@import "~bootstrap/scss/button-group";
-@import "~bootstrap/scss/input-group";
-@import "~bootstrap/scss/custom-forms";
-@import "~bootstrap/scss/breadcrumb";
-@import "~bootstrap/scss/pagination";
-@import "~bootstrap/scss/badge";
-@import "~bootstrap/scss/jumbotron";
-@import "~bootstrap/scss/alert";
-@import "~bootstrap/scss/progress";
-@import "~bootstrap/scss/media";
-@import "~bootstrap/scss/list-group";
-@import "~bootstrap/scss/close";
-@import "~bootstrap/scss/modal";
-@import "~bootstrap/scss/tooltip";
-@import "~bootstrap/scss/popover";
-@import "~bootstrap/scss/carousel";
-@import "~bootstrap/scss/utilities";
-
-/*------------------------------------*\
-    Patterns
-\*------------------------------------*/
-
-// Atoms
-@import "../../patterns/00-atoms/spinner/spinner";
-@import "../../patterns/00-atoms/button/button";
-@import "../../patterns/00-atoms/icons/icons";
-
-// Molecules
-@import "../../patterns/01-molecules/breadcrumb/breadcrumb";
-@import "../../patterns/01-molecules/card/card";
-@import "../../patterns/01-molecules/dropdown/dropdown";
-@import "../../patterns/01-molecules/forms/form-group";
-@import "../../patterns/01-molecules/forms/select";
-@import "../../patterns/01-molecules/input-group/input-group";
-@import "../../patterns/01-molecules/modal/modal";
-@import "../../patterns/01-molecules/navbar/navbar";
-@import "../../patterns/01-molecules/navbar/navbar-form";
-@import "../../patterns/01-molecules/messages/notification/index";
-@import "../../patterns/01-molecules/messages/banner-message/banner-message";
-@import "../../patterns/01-molecules/page-header/page-header";
-@import "../../patterns/01-molecules/progress-tracker/progress-tracker";
-@import "../../patterns/01-molecules/website/usp";
-
-// Organisms
-@import "../../patterns/02-organisms/auth/login";
-@import "../../patterns/02-organisms/layout/footer";
-@import "../../patterns/02-organisms/navigation/nav-secondary/nav-secondary";
-@import "../../patterns/02-organisms/navigation/nav-sidebar";
-@import "../../patterns/02-organisms/slider/references-slider";
-@import "../../patterns/02-organisms/tabs/vertical-tabs";
