+++ source/pages/product/index.js
@@ -1,0 +1,15 @@
+import * as Api from './api';
+import * as Apps from './apps';
+import * as Widgets from './widgets';
+import * as BigData from './big-data';
+import * as Platform from './platform';
+import * as Sdk from './sdk';
+
+export {
+    Api,
+    Apps,
+    Widgets,
+    BigData,
+    Platform,
+    Sdk
+}
